requirements.txt
attrs==23.2.0
certifi==2024.7.4
exceptiongroup==1.2.2
Faker==26.0.0
greenlet==3.0.3
h11==0.14.0
idna==3.7
outcome==1.3.0.post0
psycopg2-binary==2.9.9
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
selenium==4.22.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.31
trio==0.26.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.2
websocket-client==1.8.0
wsproto==1.2.0
fake-useragent==1.5.1
beautifulsoup4==4.12.3

